kaspi-saas/backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
sqlalchemy==2.0.30
alembic==1.13.1
psycopg2-binary==2.9.9
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.9
celery==5.4.0
redis==5.0.4
requests==2.31.0
beautifulsoup4==4.12.3
lxml-html-clean
schedule==1.2.1
python-dotenv==1.0.1
httpx==0.27.0
pydantic[email]==2.7.1
